refactor pump moves to support parallelism

+++ move_pump_runs2opendata_bucket.sh
@@ -1,8 +1,10 @@
 #!/usr/bin/env bash
 set -exo pipefail
 #assume the AWS Open Data Creds are available and are being used for these operations
+export LC_ALL=C
 export LOG_DIR=../transfer_logs
 mkdir -p $LOG_DIR
+export threads=10
 #in seconds
 SLEEP_TIME=60
 #source bucket
@@ -22,12 +24,15 @@ pSB=$SB/pump-outputs/$org0
 pTB=$TB/recount3expansion/pump/$org0
 pDONES="s3://monorail-batch/PUMP_DONES"
 
+#REFACTORED to do parallel copyies/removals
+
 while true; do
     #remove any previous done files to avoid confusion about what's already been copied
     rm -rf *.DONE
-    #aws s3 ls $pDONES | fgrep ".DONE" | tr -s " " $'\t' | cut -f 4 > pdones
     #get latest set of finsihed samples (from pump)
     aws s3 --profile $awsprofile cp --request-payer requester --recursive $pDONES/ ./
+    echo -n "" > $LOG_DIR/all_copies.${date}
+    echo -n "" > copy_finished_pump_runs2opendata.jobs
     #for each sample (.DONE file):
     for f in `ls *.DONE`; do
         #get the full S3 source path to copy the pump results from
@@ -37,11 +42,45 @@ while true; do
         logd=$LOG_DIR/$suffix
         mkdir -p $logd
         #copy back the full pump results directory to the target bucket
-        /usr/bin/time -v aws s3 --profile $awsprofile sync --request-payer requester $s3path/ $pTB/$suffix/ > $logd/s3copy 2>&1
-        /usr/bin/time -v aws s3 --profile $awsprofile cp --request-payer requester $pDONES/$f $pTB/DONES/ > $logd/s3copy.done 2>&1
-        #then remove path from source bucket
-        /usr/bin/time -v aws s3 --profile $awsprofile rm --request-payer requester --recursive $s3path/ > $logd/s3rm 2>&1
-        /usr/bin/time -v aws s3 --profile $awsprofile rm --request-payer requester $pDONES/$f > $logd/s3rm.done 2>&1
+        echo "/usr/bin/time -v aws s3 --profile $awsprofile sync --request-payer requester $s3path/ $pTB/$suffix/ > $logd/s3copy 2>&1" >> copy_finished_pump_runs2opendata.jobs
+        echo "$f|$s3path|$suffix" >> $LOG_DIR/all_copies.${date}
+    done
+    /usr/bin/time -v parallel -j$threads < copy_finished_pump_runs2opendata.jobs > copy_finished_pump_runs2opendata.jobs.run${threads} 2>&1
+    echo -n "" > $LOG_DIR/bad_copies.${date}
+    echo -n "" > $LOG_DIR/all_removals.${date}
+    echo -n "" > remove_finished_pump_runs2opendata.jobs
+    for copyline in `cat $LOG_DIR/all_copies.${date}`; do
+        f=$(echo "$copyline" | cut -d'|' -f 1)
+        s3path=$(echo "$copyline" | cut -d'|' -f 2)
+        suffix=$(echo "$copyline" | cut -d'|' -f 3)
+        logd=$LOG_DIR/$suffix
+        set +eo pipefail
+        good=$(fgrep "Exit status: 0" $logd/s3copy)
+        set -eo pipefail
+        if [[ -n $good ]]; then
+            /usr/bin/time -v aws s3 --profile $awsprofile cp --request-payer requester $pDONES/$f $pTB/DONES/ > $logd/s3copy.done 2>&1
+            #then remove path from source bucket
+            echo "/usr/bin/time -v aws s3 --profile $awsprofile rm --request-payer requester --recursive $s3path/ > $logd/s3rm 2>&1" >> remove_finished_pump_runs2opendata.jobs
+        echo "$f|$s3path|$suffix" >> $LOG_DIR/all_removals.${date}
+        else
+            echo "$copyline" >> $LOG_DIR/bad_copies.${date}
+        fi
+    done
+    /usr/bin/time -v parallel -j$threads < remove_finished_pump_runs2opendata.jobs > remove_finished_pump_runs2opendata.jobs${threads} 2>&1
+    echo -n "" > $LOG_DIR/bad_removals.${date}
+    for copyline in `cat $LOG_DIR/all_removals.${date}`; do
+        f=$(echo "$copyline" | cut -d'|' -f 1)
+        s3path=$(echo "$copyline" | cut -d'|' -f 2)
+        suffix=$(echo "$copyline" | cut -d'|' -f 3)
+        logd=$LOG_DIR/$suffix
+        set +eo pipefail
+        good=$(fgrep "Exit status: 0" $logd/s3rm)
+        set -eo pipefail
+        if [[ -n $good ]]; then
+            /usr/bin/time -v aws s3 --profile $awsprofile rm --request-payer requester $pDONES/$f > $logd/s3rm.done 2>&1
+        else
+            echo "$copyline" >> $LOG_DIR/bad_removals.${date}
+        fi
     done
     sleep $SLEEP_TIME
 done
